# refactor: improve generate_build_file.ts performance