--- a/internal/npm_install/tsconfig.json
+++ b/internal/npm_install/tsconfig.json
@@ -1,11 +1,11 @@
 {
   "compilerOptions": {
-    "lib": ["es2017", "dom"],
+    "lib": ["es2017" , "es2018"],
     "module": "commonjs",
     "removeComments": true,
     "skipLibCheck": true,
     "strict": true,
-    "target": "es2015"
+    "target": "es2018"
   },
   "files": [
     "generate_build_file.ts",