… internal/npm_install/test/generate_build_file.spec.js
… internal/npm_install/test/golden/BUILD.bazel.golden
… internal/npm_install/test/golden_directory_artifacts/BUILD.bazel.golden
… internal/npm_install/test/golden_multi_linked/BUILD.bazel.golden