@cuyl 2021-09-27
> There's a lot of random missing semicolons throughout the file
@jbedard
A lot of random missing before as well, I fixed it all. maybe we could add an eslint rule to check `ts,js` code styles.