@jbedard 2021-09-27
> maybe we could add an eslint rule to check ts,js code styles.

I agree with that one, would lead to less confusion on my part in PRs like this :)